Move PlannedActualTable out of shared.js
